# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareGrid AI - District Healthcare Operations Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: centerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
...
  /inventory/scan:
    post:
      operationId: scanInventory
      tags: [inventory]
      summary: Simulate OCR scan to update medicine stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found

  # Daily Logs
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: Get daily patient/staff logs
      parameters:
        - name: centerId
          in: query
          required: false
          schema:
            type: integer
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Daily logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
    post:
      operationId: createLog
      tags: [logs]
      summary: Create or update a daily log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogInput"
      responses:
        "201":
          description: Created log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
  /logs/voice:
    post:
      operationId: processVoiceLog
      tags: [logs]
      summary: Process a voice log entry via AI parsing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceLogInput"
      responses:
        "200":
          description: Parsed and applied voice log result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceLogResult"

  # Attendance
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [id, centerId, centerName, medicineName, quantity, dailyBurnRate, status]
      properties:
        id:
          type: integer
        centerId:
          type: integer
        centerName:
          type: string
        medicineName:
          type: string
        quantity:
          type: integer
        dailyBurnRate:
          type: number
        status:
          type: string
          enum: [ok, low, critical, surplus]
        daysRemaining:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    InventoryUpdate:
      type: object
      properties:
        quantity:
          type: integer
        dailyBurnRate:
          type: number
    ScanInput:
      type: object
      required: [centerId, imageText]
      properties:
        centerId:
          type: integer
        imageText:
          type: string
        quantity:
          type: integer
...
    DailyLog:
      type: object
      required: [id, centerId, centerName, date, patientFootfall, doctorsPresent, aiAnomalyFlag]
      properties:
        id:
          type: integer
        centerId:
          type: integer
        centerName:
          type: string
        date:
          type: string
        patientFootfall:
          type: integer
        doctorsPresent:
          type: integer
        aiAnomalyFlag:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LogInput:
      type: object
      required: [centerId, date, patientFootfall, doctorsPresent]
      properties:
        centerId:
          type: integer
        date:
          type: string
        patientFootfall:
          type: integer
        doctorsPresent:
          type: integer
        aiAnomalyFlag:
          type: boolean
        notes:
          type: string
    VoiceLogInput:
      type: object
      required: [centerId, transcript]
      properties:
        centerId:
          type: integer
        transcript:
          type: string
        language:
          type: string
    VoiceLogResult:
      type: object
      required: [parsed, log, message]
      properties:
        parsed:
          type: object
          properties:
            patientFootfall:
              type: ["integer", "null"]
            medicineName:
              type: ["string", "null"]
            stockStatus:
              type: ["string", "null"]
        log:
          $ref: "#/components/schemas/DailyLog"
        message:
          type: string
...