# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CareGrid AI - District Healthcare Operations Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanInput {
  centerId: number;
  imageText: string;
  quantity?: number;
}